Fixed physical memory manager, and temporarily disabled paging

+++ src/include/lib/bitmap.h
@@ -3,8 +3,8 @@
 
 #include <stdint.h>
 
-void bitmap_set(uint8_t * bitmap, int n);
-void bitmap_unset(uint8_t * bitmap, int n);
-int bitmap_test(uint8_t * bitmap, int n);
+void bitmap_set(uint32_t * bitmap, int n);
+void bitmap_unset(uint32_t * bitmap, int n);
+int bitmap_test(uint32_t * bitmap, int n);
 
 #endif
